Add audio queue
